2.1/3

+++ code/game/objects/items/weapons/grenades/syndieminibomb.dm
@@ -5,8 +5,15 @@
 	icon_state = "syndicate"
 	item_state = "flashbang"
 	origin_tech = "materials=3;magnets=4;syndicate=2"
+	var/defused = FALSE // Used in `/obj/item/pen/fancy/bomb`.
 
 /obj/item/grenade/syndieminibomb/prime()
-	update_mob()
-	explosion(loc, 1, 2, 4, flame_range = 2, cause = src)
-	qdel(src)
+	if(defused)
+		defused = FALSE
+		if(istype(loc, /obj/item/pen/fancy/bomb))
+			var/obj/item/pen/fancy/bomb/pen_holder = loc
+			pen_holder.clickscount = 0
+	else
+		update_mob()
+		explosion(loc, 1, 2, 4, flame_range = 2, cause = src)
+		qdel(src)
